Imroving on events

+++ app/Listeners/AddCompletedDevice.php
@@ -24,29 +24,21 @@ class AddCompletedDevice
      */
     public function handle(DeleteDevice $event): void
     {
-        $device = Device::find($event->id);
-        if ($device) {
-            if ($device->deliver_to_client === true) {
-                $client = $device->client;
-                if ($client) {
-                    $completedDevice = CompletedDevice::create([
-                        'model' => $device->model,
-                        'imei' => $device->imei,
-                        'client_id' => $device->client_id,
-                        'user_id' => $device->user_id,
-                        'info' => $device->info,
-                        'problem' => $device->problem,
-                        'cost' => $device->cost,
-                        'status' => $device->status,
-                        'fix_steps' => $device->fix_steps,
-                        'date_receipt' => $device->date_receipt,
-                    ]);
-                    if ($completedDevice) {
-                        $client->decrement('devices_count');
-                    }
-                    if ($device->deliver_to_customer === true) {
-                        $device->delete();
-                    }
+
+        $device = $event->Device;
+        if ($device->deliver_to_client === true) {
+            $client = $device->client;
+            $user = $device->user;
+            if ($client && $user) {
+                $completedDevice = $device->toArray();
+                $completedDevice['client_name'] = $client->name;
+                $completedDevice['user_name'] = $user->name;
+                $completedDevice = CompletedDevice::create($completedDevice);
+                if ($completedDevice) {
+                    $client->decrement('devices_count');
+                }
+                if ($device->deliver_to_customer === true) {
+                    $device->delete();
                 }
             }
         }
